Minor speedups based on profiling with gprof on the Amiga, running fractal_mountains; including improving SLEEP to avoid busy-polling when message ports can be used easily.

+++ src/platform_generic.c
@@ -381,7 +381,7 @@ bool PfRepresentsStdin(const char *name)
 	return strcmp(name, "-") == 0;
 }
 
-void PfSleep(unsigned micros)
+void PfSleep(unsigned long micros)
 {
 	usleep(micros);
 }
